requirements.txt
streamlit
pandas
openai
matplotlib
seaborn
google-generativeai